move schema based element addressing from MVS to the core library
